+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Additional Calypso requirement related to the transaction audit data (issue [#68]).
 ### Upgraded
 - "Calypsonet Terminal Calypso API" to version `1.2.0` (issue [#66]).
-- "Keyple Util Library" to version `2.1.0` by removing deprecated methods invocation.
+- "Keyple Util Library" to version `2.1.0`.
 ### Fixed
 - Null Pointer Exception raised when the "open secure session" command returns an unexpected status word [#70]
 - JSON serialization for interfaces in objects trees (issue [#71]).